requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aniso8601==9.0.1
    # via flask-restful
click==8.0.1
    # via flask
colorama==0.4.4
    # via click
flask-restful==0.3.9
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via -r requirements.in
flask==2.0.1
    # via
    #   -r requirements.in
    #   flask-restful
    #   flask-sqlalchemy
greenlet==1.1.0
    # via sqlalchemy
itsdangerous==2.0.1
    # via flask
jinja2==3.0.1
    # via flask
markupsafe==2.0.1
    # via jinja2
python-dotenv==0.18.0
    # via -r requirements.in
pytz==2021.1
    # via flask-restful
six==1.16.0
    # via flask-restful
sqlalchemy==1.4.19
    # via flask-sqlalchemy
werkzeug==2.0.1
    # via flask
